Refactor code into new classes and packages
